Rename RendererProvider to rendererProperties
